ai_service/CodeSprint-katya/requirements.txt
dlt[postgres]>=0.5.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
flask>=2.3.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
